Rearranging DigestCreator and DigestHolder
1) Added the proper way for type subsetting
2) Added skeleton code for proper maintainability.
