+++ util/src/DigestImpl.cpp
@@ -9,97 +9,84 @@
 // these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE
 // file.
 
-#include "DigestImpl.hpp"
-#include "hex_tools.h"
+#include <string.h>
+#include <iomanip>
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wpedantic"
+#include <cryptopp/dll.h>
+#include <cryptopp/pem.h>
+#pragma GCC diagnostic pop
+
+#include "assertUtils.hpp"
+#include <cryptopp/cryptlib.h>
+#include <cryptopp/ida.h>
+#include <cryptopp/eccrypto.h>
+
+#include "DigestImpl.ipp"
+
+#if defined MD5_DIGEST
+#include <cryptopp/md5.h>
+#define DigestType Weak1::MD5
+#elif defined SHA256_DIGEST
+#define DigestType SHA256
+#elif defined SHA512_DIGEST
+#define DigestType SHA512
+#endif
+
+using namespace CryptoPP;
 
 namespace concord::util::digest {
 
-template <class T>
-DigestHolder<T>::DigestHolder() {
-  d.fill(0);
-};
+size_t DigestUtil::digestLength() { return DigestType::DIGESTSIZE; }
 
-template <class T>
-DigestHolder<T>::DigestHolder(unsigned char initVal) {
-  d.fill(initVal);
-}
+bool DigestUtil::compute(const char* input,
+                         size_t inputLength,
+                         char* outBufferForDigest,
+                         size_t lengthOfBufferForDigest) {
+  DigestType dig;
 
-template <class T>
-DigestHolder<T>::DigestHolder(const char* other) {
-  memcpy(d.data(), other, DIGEST_SIZE);
-}
+  size_t size = dig.DigestSize();
 
-template <class T>
-DigestHolder<T>::DigestHolder(char* buf, size_t len) {}
+  if (lengthOfBufferForDigest < size) return false;
 
-template <class T>
-char* DigestHolder<T>::content() const {
-  return (char*)d.begin();
-}
+  SecByteBlock digest(size);
 
-template <class T>
-void DigestHolder<T>::makeZero() {
-  d.fill(0);
-}
+  dig.Update((CryptoPP::byte*)input, inputLength);
+  dig.Final(digest);
+  const CryptoPP::byte* h = digest;
+  memcpy(outBufferForDigest, h, size);
 
-template <class T>
-std::string DigestHolder<T>::toString() const {
-  return concordUtils::bufferToHex(d, DIGEST_SIZE, false);
-}
-
-template <class T>
-void DigestHolder<T>::print() {
-  printf("digest=[%s]", toString().c_str());
-}
-
-template <class T>
-const char* const DigestHolder<T>::get() const {
-  return (const char*)d.data();
-}
-
-template <class T>
-char* DigestHolder<T>::getForUpdate() {
-  return (char*)d.data();
-}
-
-template <class T>
-bool DigestHolder<T>::isZero() const {
-  for (int i = 0; i < DIGEST_SIZE; ++i) {
-    if (d[i] != 0) return false;
-  }
   return true;
 }
 
-template <class T>
-int DigestHolder<T>::hash() const {
-  uint64_t* p = (uint64_t*)d.begin();
-  int h = (int)p[0];
-  return h;
+DigestUtil::Context::Context() {
+  DigestType* p = new DigestType();
+  internalState = p;
 }
 
-template <class T>
-bool DigestHolder<T>::operator==(const DigestHolder& other) const {
-  int r = memcmp(d, other.d, DIGEST_SIZE);
-  return (r == 0);
+void DigestUtil::Context::update(const char* data, size_t len) {
+  ConcordAssert(internalState != NULL);
+  DigestType* p = (DigestType*)internalState;
+  p->Update((CryptoPP::byte*)data, len);
 }
 
-template <class T>
-bool DigestHolder<T>::operator!=(const DigestHolder& other) const {
-  int r = memcmp(d, other.d, DIGEST_SIZE);
-  return (r != 0);
+void DigestUtil::Context::writeDigest(char* outDigest) {
+  ConcordAssert(internalState != NULL);
+  DigestType* p = (DigestType*)internalState;
+  SecByteBlock digest(digestLength());
+  p->Final(digest);
+  const CryptoPP::byte* h = digest;
+  memcpy(outDigest, h, digestLength());
+
+  delete p;
+  internalState = NULL;
 }
 
-template <class T>
-DigestHolder<T>& DigestHolder<T>::operator=(const DigestHolder<T>& other) {
-  if (this == &other) {
-    return *this;
+DigestUtil::Context::~Context() {
+  if (internalState != NULL) {
+    DigestType* p = (DigestType*)internalState;
+    delete p;
+    internalState = NULL;
   }
-  memcpy(d, other.d, DIGEST_SIZE);
-  return *this;
-}
-
-template <class T>
-void DigestHolder<T>::digestOfDigest(const DigestHolder& inDigest, DigestHolder& outDigest) {
-  DigestUtil::compute(inDigest.d, DIGEST_SIZE, outDigest.d, DIGEST_SIZE);
 }
 }  // namespace concord::util::digest
